+++ src/Header/Common/ContentType.php
@@ -4,8 +4,7 @@ declare(strict_types=1);
 
 namespace Stadly\Http\Header\Common;
 
-use Stadly\Http\Header\Value\MediaType;
-use Stadly\Http\Header\Value\MediaType\Parameter;
+use Stadly\Http\Header\Value\MediaType\MediaType;
 
 /**
  * Class for handling the HTTP header field Content-Type.