+++ src/Header/Value/EntityTag/EntityTag.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Stadly\Http\Header\Value;
+namespace Stadly\Http\Header\Value\EntityTag;
 
 use InvalidArgumentException;
 use Stadly\Http\Utilities\Rfc7232;