+++ tests/Header/Response/ETagTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Stadly\Http\Header\Response;
 
 use PHPUnit\Framework\TestCase;
-use Stadly\Http\Header\Value\EntityTag;
+use Stadly\Http\Header\Value\EntityTag\EntityTag;
 
 /**
  * @coversDefaultClass \Stadly\Http\Header\Response\ETag