+++ tests/Header/Response/HeaderFactoryTest.php
@@ -8,8 +8,8 @@ use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 use Stadly\Http\Header\Common\Header;
 use Stadly\Http\Header\Common\ContentType;
-use Stadly\Http\Header\Value\EntityTag;
-use Stadly\Http\Header\Value\MediaType;
+use Stadly\Http\Header\Value\EntityTag\EntityTag;
+use Stadly\Http\Header\Value\MediaType\MediaType;
 
 /**
  * @coversDefaultClass \Stadly\Http\Header\Response\HeaderFactory